Modificado Ejercicio 6, solo poner un Scanner

+++ Tema 1/Ejemplo6.java
@@ -3,7 +3,7 @@ public class Ejemplo6 {
     public static void main (String argv[]) {
         float area;
         final double PI = 3.14159;
-        int radio;
+        float radio;
         float length;
 
         System.out.println("Pide el radio de una circunferencia");
